Terms of Services
Terms of Service for Empower with Adele
Effective Date: 29/06/24

1. Acceptance of Terms
By accessing and using the Empower with Adele website and services, you accept and agree to be bound by the terms and provision of this agreement. In addition, when using these particular services, you shall be subject to any posted guidelines or rules applicable to such services. Any participation in this service will constitute acceptance of this agreement. If you do not agree to abide by the above, please do not use this service.

2. Privacy Policy
Your use of the website is also subject to our Privacy Policy. Please review our Privacy Policy, which also governs the site and informs users of our data collection practices.

3. Description of Service
Empower with Adele is a platform providing motivational and inspirational content. We strive to inspire and empower individuals through various forms of media, including videos and articles.

4. User Accounts
If you create an account on the website, you are responsible for maintaining the security of your account, and you are fully responsible for all activities that occur under the account and any other actions taken in connection with it. You must immediately notify us of any unauthorized uses of your account or any other breaches of security. We will not be liable for any acts or omissions by you, including any damages of any kind incurred as a result of such acts or omissions.

5. Use of Service
You agree not to use the Service for any purpose that is unlawful or prohibited by these Terms. You may not use the Service in any manner that could damage, disable, overburden, or impair the website or interfere with any other party's use and enjoyment of the Service.

6. Intellectual Property
The content, layout, design, data, databases, and graphics on this website are protected by international intellectual property laws and are owned by Empower with Adele or its licensors. Unless expressly permitted in writing in a license agreement, no part of the Service may be reproduced, distributed, or transmitted in any form or by any means, including photocopying, recording, or other electronic or mechanical methods.

7. Termination
We may terminate your access to the site, without cause or notice, which may result in the forfeiture and destruction of all information associated with your account. All provisions of this agreement that, by their nature, should survive termination shall survive termination, including, without limitation, ownership provisions, warranty disclaimers, indemnity, and limitations of liability.

8. Limitation of Liability
In no event shall Empower with Adele, nor its directors, employees, partners, agents, suppliers, or affiliates, be liable for any indirect, incidental, special, consequential, or punitive damages, including without limitation, loss of profits, data, use, goodwill, or other intangible losses, resulting from (i) your access to or use of or inability to access or use the service; (ii) any conduct or content of any third party on the service; (iii) any content obtained from the service; and (iv) unauthorized access, use, or alteration of your transmissions or content, whether based on warranty, contract, tort (including negligence) or any other legal theory, whether or not we have been informed of the possibility of such damage, and even if a remedy set forth herein is found to have failed of its essential purpose.

9. Governing Law
These Terms shall be governed and construed in accordance with the laws of New Zealand, without regard to its conflict of law provisions.

10. Changes to Terms
We reserve the right, at our sole discretion, to modify or replace these Terms at any time. If a revision is material, we will try to provide at least 30 days notice prior to any new terms taking effect. What constitutes a material change will be determined at our sole discretion.

11. Contact Us
If you have any questions about these Terms, please contact us at empowerwithadele@gmail.com